fftools/ffmpeg: Added ability to set a input burst time before readrate is enforced
Implemented is the -irb <seconds> flag (i.e. "initial read burst") that
causes ffmpeg to read the specified number of seconds of input before a
given readrate starts to be enforced.  All inputs have to reach this
point before the readrate is enforced.  The reason for this change is,
in the scenario of live-streaming by transcoding from an pre-recorded or
delayed source to an HLS or DASH webroot, we can more quickly prime the
system of the first few seconds before the real-world read rate is
imposed.  Else, one must wait for an entire segment length before the
data can be fetched from the web server.

Signed-off-by: Davy Durham <ddurham@users.sourceforge.net>