fftools/ffmpeg: sleep more efficiently when using -re or -readrate
When the ffmpeg cli tool is given a read rate (i.e. -re or -readrate), the read
function is called repeatedly but which may return EAGAIN if it should be called
again later.  How much later is not indicated and the trancode loop just sleeps
for 10ms and tries again.  On the next try it may get EGAIN again and this adds
up to wasted CPU.  For example, conversions of audio files show about a 15-20%
higher CPU usage than necessary, when compared to running without a read rate.
This patch introduces InputFile::next_read_time which is computed on the way out
of the read function, ifile_get_packet(), when it returns EAGAIN.  Before
sleeping, the transcoding loop now consults this value to determine how long it
should sleep rather than blindly sleeping 10ms.
Running ffmpeg under the 'time' command shows an improved real+sys time after
this change, and these savings improve further if one specifies a larger pkt
size to the demuxer (e.g. -max_size for .wav, -raw_packet_size for raw, etc.)
since the system is able to sleep longer (and now accurately) each iteration.

Signed-off-by: Davy Durham <ddurham@users.sourceforge.net>